fix: improve docker container
