+++ Dockerfile
@@ -1,5 +1,33 @@
 # syntax=docker/dockerfile:1
-FROM docker.io/golang:1.22-alpine
+# Use the official golang image to create a build artifact.
+FROM docker.io/golang:1.22-alpine as builder
+WORKDIR /src
+RUN mkdir app
+
+# Setup dependencies
+RUN apk add --no-cache nodejs npm
+# Install templ
+RUN go install github.com/a-h/templ/cmd/templ@latest
+
+# Copy source (see .dockerignore for exclusions)
+COPY . .
+
+# Install Frontend
+RUN npx pnpm install --frozen-lockfile
+# "Build" the frontend files
+RUN npx pnpm build
+# Compile the templ files
+RUN templ compile
+# Build the Go app
+RUN go build -o /app/main ./cmd/main.go
+
+# Copy the frontend files to the app
+RUN cp -r assets /app/assets
+
+# Use a smaller image to run the app
+FROM docker.io/alpine:latest
+COPY --from=builder /app/main /app/main
+COPY --from=builder /app/assets /app/assets
 
 LABEL maintainer="Nicholas Westerhausen <nicholas@mail.nwest.one>"
 LABEL description="A simple domain monitoring tool, which tracts the expiration date of domains and sends notifications."
@@ -7,33 +35,8 @@ LABEL description="A simple domain monitoring tool, which tracts the expiration 
 WORKDIR /app
 
 # Setup the data directory
-RUN mkdir -p /app/data
+RUN mkdir /app/data
 VOLUME /app/data
-
-# Install nodejs
-RUN apk add --no-cache nodejs npm
-
-# Copy go mod and sum files to the working directory
-COPY go.mod go.sum ./
-# Download all dependencies. Dependencies will be cached if the go.mod and go.sum files are not changed
-RUN go mod download
-
-# Install pnpm
-RUN npm install -g pnpm
-# Copy the pnpm-lock.yaml and package.json files to the working directory
-COPY pnpm-lock.yaml package.json ./
-# Install all the dependencies
-RUN pnpm install
-
-# Copy the asset files
-COPY assets ./assets
-# Copy the source from the current directory to the Working Directory inside the container
-COPY . ./
-
-# Build the Go app
-RUN go build ./cmd/main.go
-# Build the frontend dependencies
-RUN pnpm build
 
 # Expose the default port
 EXPOSE 3124
