+++ README.md
@@ -54,11 +54,11 @@ services:
   dm:
     image: ghcr.io/nwesterhausen/domain-monitor/server:latest
     ports:
-    - 127.0.0.1:4201:4201/tcp
+      - 127.0.0.1:4201:4201/tcp
     volumes:
-    - ./config:/app/config:rw
-    - ./whois:/app/whois-data:rw
-version: '3.9'
+      - ./config:/app/config:rw
+      - ./whois:/app/whois-data:rw
+version: "3.9"
 ```
 
 #### Using [Docker Hub](https://hub.docker.com/repository/docker/nwesterhausen/domain-monitor)
@@ -149,7 +149,6 @@ monitor. Each domain has the following properties:
 | enabled     | bool   | If true, whois lookups will be done (on the schedule described above) for this domain.                                    |
 | (_todo_) id | string | A generated ID string. If you are manually adding to domain.yaml do not add an ID, it will be generated by domain-monitor |
 
-
 ## Development
 
 - install go, node, pnpm
@@ -157,5 +156,7 @@ monitor. Each domain has the following properties:
 - install templ `go install github.com/a-h/templ/cmd/templ@latest`
 
 In root of repository do `pnpm i` to install node dependencies
+
+To compile the templates, run `templ compile` in the root of the repository.
 
 Run dev server with `air`
