+++ handlers/routes.go
@@ -4,7 +4,6 @@ import (
 	"github.com/a-h/templ"
 	"github.com/labstack/echo/v4"
 	"github.com/nwesterhausen/domain-monitor/configuration"
-	"github.com/nwesterhausen/domain-monitor/mailer"
 	"github.com/nwesterhausen/domain-monitor/service"
 )
 
@@ -43,7 +42,7 @@ func SetupConfigRoutes(app *echo.Echo, config configuration.Configuration) {
 	// configGroup.PUT("/", ch.HandleConfigUpdate)
 }
 
-func SetupMailerRoutes(app *echo.Echo, ms *mailer.MailerService, alertRecipient string) {
+func SetupMailerRoutes(app *echo.Echo, ms *service.MailerService, alertRecipient string) {
 	mailerGroup := app.Group("/mailer")
 
 	mh := NewMailerHandler(ms, alertRecipient)
